+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2018-11-30
+
+- Updated email verification language.
+- Optimized wish list gift sorting.
+
+# 2018-11-28
+
+- Fixed totals for dibs page.
+
 # 2018-11-21
 
 - Initial release.